feat(TypeHandlerLibrary): split TypeHandlerLibraryTest and setup default handlers correctly
