+++ engine-tests/src/test/java/org/terasology/persistence/typeHandling/TypeHandlerLibraryImplTest.java
@@ -1,41 +1,28 @@
 // Copyright 2020 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
+
 package org.terasology.persistence.typeHandling;
 
 import org.junit.jupiter.api.Test;
 import org.reflections.Reflections;
 import org.terasology.persistence.typeHandling.coreTypes.CollectionTypeHandler;
-import org.terasology.persistence.typeHandling.coreTypes.EnumTypeHandler;
 import org.terasology.persistence.typeHandling.coreTypes.ObjectFieldMapTypeHandler;
-import org.terasology.persistence.typeHandling.coreTypes.RuntimeDelegatingTypeHandler;
-import org.terasology.persistence.typeHandling.coreTypes.StringMapTypeHandler;
 import org.terasology.reflection.MappedContainer;
 import org.terasology.reflection.TypeInfo;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import java.util.Queue;
 import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class TypeHandlerLibraryTest {
+class TypeHandlerLibraryImplTest {
     private final Reflections reflections = new Reflections(getClass().getClassLoader());
     private final TypeHandlerLibrary typeHandlerLibrary = new TypeHandlerLibraryImpl(reflections);
 
-    private enum AnEnum {}
 
     @MappedContainer
     private static class AMappedContainer {}
-
-    @Test
-    public void testEnumHandler() {
-        TypeHandler<AnEnum> handler = typeHandlerLibrary.getTypeHandler(AnEnum.class).get();
-
-        assertTrue(handler instanceof EnumTypeHandler);
-    }
 
     @Test
     public void testMappedContainerHandler() {
@@ -62,26 +49,4 @@ public class TypeHandlerLibraryTest {
         assertTrue(queueHandler instanceof CollectionTypeHandler);
     }
 
-    @Test
-    public void testStringMapHandler() {
-        TypeHandler<Map<String, Integer>> handler =
-                typeHandlerLibrary.getTypeHandler(new TypeInfo<Map<String, Integer>>() {}).get();
-
-        assertTrue(handler instanceof StringMapTypeHandler);
-    }
-
-    @Test
-    public void testInvalidTypeHandler() {
-        Optional<TypeHandler<Map<Integer, Integer>>> handler =
-                typeHandlerLibrary.getTypeHandler(new TypeInfo<Map<Integer, Integer>>() {});
-
-        assertFalse(handler.isPresent());
-    }
-
-    @Test
-    public void testGetBaseTypeHandler() {
-        TypeHandler<Integer> handler = typeHandlerLibrary.getBaseTypeHandler(TypeInfo.of(Integer.class));
-
-        assertTrue(handler instanceof RuntimeDelegatingTypeHandler);
-    }
 }
